requirements.txt
Flask
pymongo
dnspython
PyPDF2
python-docx
scikit-learn
